refactor: abstract configs as a domain level object
